Prevent int/byte conversion.
git-svn-id: http://svn.codehaus.org/jruby/joni/trunk@6892 961051c9-f516-0410-bf72-c9f7e237a7b7

+++ src/org/joni/encoding/Encoding.java
@@ -404,11 +404,11 @@ public abstract class Encoding {
         return code < 128; 
     }
     
-    public static int asciiToLower(int c) {
+    public static byte asciiToLower(int c) {
         return AbstractEncoding.AsciiToLowerCaseTable[c];
     }
     
-    public static int asciiToUpper(int c) {
+    public static byte asciiToUpper(int c) {
         return AbstractEncoding.AsciiToUpperCaseTable[c];
     }
